+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - unified custom features
 - money management for cosmere/nimble characters
 - conditions management for cosmere characters
+- expertise editing for cosmere characters
 
 ### Modified
 - leveling up safe mode